chore/client: rename file import
