+++ client/src/App.tsx
@@ -12,7 +12,7 @@ import WalletApi, {
     CIP0030Wallet,
     WalletKeys,
 } from "./services/connectors/wallet.connector";
-import { getNetworkId } from "./services/http.services";
+import { getNetworkId } from "./services/claim.services";
 import "./styles.scss";
 
 export const Themes = {
