+++ client/src/pages/Rewards/components/DepositInfo/index.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import TransactionDetail from "../TransactionDetail";
 import { formatTokens } from "src/services/utils.services";
 import { useEffect, useState } from "react";
-import { getTxStatus } from "src/services/http.services";
+import { getTxStatus } from "src/services/claim.services";
 import { GetCustomRewards } from "src/entities/vm.entities";
 import WalletApi from "src/services/connectors/wallet.connector";
 import SendAdaInfo from "../SendAdaInfo";
